requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
anyio==4.4.0
apache-airflow==2.9.2
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.0
apache-airflow-providers-fab==1.1.1
apache-airflow-providers-ftp==3.9.1
apache-airflow-providers-http==4.11.1
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-snowflake==5.5.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
arch==7.0.0
argcomplete==3.4.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
clarabel==0.7.1
click==8.1.7
clickclick==20.10.2
cliq==0.9.3
cloudpickle==3.0.0
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.2.1
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.8
cvxpy==1.4.3
cycler==0.12.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docker==7.1.0
docutils==0.21.2
ecos==2.0.13
email_validator==2.1.2
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.15.3
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.51.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-speech==2.26.0
google-cloud-translate==3.15.3
google-re2==1.1.20240601
googleapis-common-protos==1.63.1
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imbalanced-learn==0.12.3
immutabledict==4.2.0
importlib-metadata==7.0.0
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
limits==3.12.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.2.2
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.8.4
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
missingpy==0.2.0
mlflow==2.13.0
more-itertools==10.3.0
multidict==6.0.5
multitasking==0.0.11
openpyxl==3.1.3
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
optbinning==0.19.0
optuna==3.6.1
ordered-set==4.1.0
ortools==9.9.3963
pandas==2.2.2
pathspec==0.12.1
patsy==0.5.6
peewee==3.17.5
pendulum==3.0.0
pillow==10.3.0
pluggy==1.5.0
prison==0.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.12.0
pycparser==2.22
pydub==0.25.1
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pytest==8.2.0
python-box==7.2.0
python-daemon==3.0.1
python-nvd3==0.16.0
python-pptx==0.6.23
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
querystring-parser==1.2.4
referencing==0.35.1
requests==2.32.2
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
ropwr==1.0.0
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.4.2
scs==3.2.4.post1
seaborn==0.13.2
setproctitle==1.3.3
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
statsmodels==0.14.2
tabulate==0.9.0
tenacity==8.4.1
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.14.1
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
XlsxWriter==3.2.0
yarl==1.9.4
